requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.8=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.6.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.9.24=haa95532_0
expat=2.6.3=h5da7b33_0
frozenlist=1.4.1=pypi_0
google-ai-generativelanguage=0.6.6=pypi_0
google-api-core=2.20.0=pypi_0
google-api-python-client=2.147.0=pypi_0
google-auth=2.35.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.7.2=pypi_0
googleapis-common-protos=1.65.0=pypi_0
grpcio=1.66.2=pypi_0
grpcio-status=1.62.3=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.2=pypi_0
langchain-core=0.3.8=pypi_0
langchain-google-genai=2.0.0=pypi_0
langchain-text-splitters=0.3.0=pypi_0
langsmith=0.1.131=pypi_0
libffi=3.4.4=hd77b12b_1
multidict=6.1.0=pypi_0
openssl=3.0.15=h827c3e9_0
orjson=3.10.7=pypi_0
pip=24.2=py312haa95532_0
proto-plus=1.24.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
python=3.12.6=h14ffc60_1
python-dotenv=1.0.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
rsa=4.9=pypi_0
setuptools=75.1.0=py312haa95532_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
yarl=1.13.1=pypi_0
zlib=1.2.13=h8cc25b3_1